HW03/code/requirements.txt
matplotlib>=3.4.3
numpy>=1.19.5
opencv_python>=4.5.4.58
Pillow>=9.0.1
